+++ CHANGELOG.md
@@ -66,4 +66,4 @@
 * **BC** - function `Psl\Password\algorithms()` have been removed.
 * **BC** - `Psl\Result\ResultInterface::getException()` method has been renamed to `Psl\Result\ResultInterface::getThrowable()`
 * **BC** - `Psl\Result\wrap` function now catches all `Throwable`s instead of only `Exception`s
-
+* introduced a new `Psl\Result\reflect` function